refactor: consolidate periods module types
